chore(cleanup): tree hygiene + verification-token typo rename
- .gitignore: ignore .idea/ and .playwright-mcp/
- git rm --cached .idea/* (5 IDE config files)
- delete fix-dev-server.bat and start-dev.ps1 (Windows-only crutches
  for a Turbopack/watcher issue; should fix root cause not workaround)
- rename verificiation-token.ts → verification-token.ts and update
  two importers (auth/tokens.ts, auth/verification/action.ts)

Skipped the audit's "two auth.ts files" dedup — investigation showed
they serve different roles (root NextAuth instance vs slim helper
wrapper) so kept both.

Closes #35
