+++ .gitignore
@@ -23,6 +23,8 @@
 # misc
 .DS_Store
 *.pem
+.idea/
+.playwright-mcp/
 
 # debug
 npm-debug.log*
